Move ScopePermissionService to its own module
